requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Security and testing
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0

# Web framework for demos
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector databases
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Machine learning
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0

# Security tools
cryptography>=41.0.0
pyjwt>=2.8.0
python-dotenv>=1.0.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
rich>=13.6.0
colorama>=0.4.6

# Docker support
docker>=6.1.0

# API testing
httpx>=0.25.0
aiohttp>=3.9.0

# Jupyter for interactive demos
jupyter>=1.0.0
ipykernel>=6.25.0 